# internal/admin/server.go
package admin

import (
	"crypto/subtle"
	"encoding/json"
	"errors"
	"fmt"
	"io/fs"
	"net"
	"net/http"
	"strconv"
	"strings"
	"sync"

	"ipv6-proxy-pool/internal/config"
	"ipv6-proxy-pool/internal/lease"
	"ipv6-proxy-pool/internal/listener"
)

// Pool describes the lease operations exposed by the management API.
type Pool interface {
	Acquire(id string, persistent bool) (lease.Lease, error)
	AcquirePort(id string, port int, persistent bool) (lease.Lease, error)
	Get(id string) (lease.Lease, bool)
	List() []lease.Lease
	ListAll() []lease.Lease
	StandbyCount() int
	Release(id string) bool
	ReleaseIdle() int
	Rotate(id string) (lease.Lease, error)
	Recycle(id string) (lease.Lease, error)
	SetPersistent(id string, persistent bool) (lease.Lease, error)
}

// Options configures management endpoints, per-IPv6 listeners, and optional
// same-origin web assets.
type Options struct {
	ConfigPath      string
	RuntimeConfig   config.Config
	AdminToken      string
	Web             fs.FS
	ListenerManager *listener.Manager
}

// Handler preserves the original lease-only API constructor.
func Handler(pool Pool) http.Handler {
	return HandlerWithOptions(pool, Options{})
}

// HandlerWithOptions returns the management API and optional web UI handler.
func HandlerWithOptions(pool Pool, options Options) http.Handler {
	server := &server{pool: pool, options: options}
	mux := http.NewServeMux()
	mux.HandleFunc("GET /healthz", server.health)
	mux.HandleFunc("GET /v1/status", server.status)
	mux.HandleFunc("GET /v1/config", server.getConfig)
	mux.HandleFunc("PUT /v1/config", server.saveConfig)
	mux.HandleFunc("GET /v1/leases", server.listLeases)
	mux.HandleFunc("POST /v1/leases", server.createLease)
	mux.HandleFunc("GET /v1/leases/{id}", server.getLease)
	mux.HandleFunc("PATCH /v1/leases/{id}", server.updateLease)
	mux.HandleFunc("DELETE /v1/leases/{id}", server.deleteLease)
	mux.HandleFunc("POST /v1/leases/{id}/rotate", server.rotateLease)
	mux.HandleFunc("POST /v1/leases/{id}/recycle", server.recycleLease)
	mux.HandleFunc("POST /v1/leases:release-idle", server.releaseIdle)
	if options.Web != nil {
		mux.Handle("/", http.FileServer(http.FS(options.Web)))
	}

	var handler http.Handler = mux
	if options.AdminToken != "" {
		handler = server.requireToken(handler)
	}
	return handler
}

// requireToken guards every /v1/* endpoint with a constant-time comparison
// against the configured admin token. Health checks and static web assets stay
// open so monitoring and the local UI remain reachable.
func (s *server) requireToken(next http.Handler) http.Handler {
	token := "Bearer " + s.options.AdminToken
	return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		if !strings.HasPrefix(r.URL.Path, "/v1/") {
			next.ServeHTTP(w, r)
			return
		}
		auth := r.Header.Get("Authorization")
		if subtle.ConstantTimeCompare([]byte(auth), []byte(token)) != 1 {
			writeError(w, http.StatusUnauthorized, errors.New("missing or invalid admin token"))
			return
		}
		next.ServeHTTP(w, r)
	})
}

type server struct {
	pool    Pool
	options Options
	mu      sync.Mutex
}

type createLeaseRequest struct {
	ID         string `json:"id"`
	Persistent bool   `json:"persistent"`
}

type updateLeaseRequest struct {
	Persistent *bool `json:"persistent"`
}

func (s *server) health(w http.ResponseWriter, _ *http.Request) {
	writeJSON(w, http.StatusOK, map[string]string{"status": "ok"})
}

func (s *server) status(w http.ResponseWriter, _ *http.Request) {
	leases := s.pool.List()
	persistent := 0
	for _, item := range leases {
		if item.Persistent {
			persistent++
		}
	}
	standby := s.pool.StandbyCount()
	writeJSON(w, http.StatusOK, map[string]any{
		"status":               "ok",
		"lease_count":          len(leases),
		"persistent_count":     persistent,
		"standby_count":        standby,
		"min_leases":           s.options.RuntimeConfig.MinLeases,
		"max_leases":           s.options.RuntimeConfig.MaxLeases,
		"ipv6_prefix":          s.options.RuntimeConfig.IPv6Prefix,
		"socks_mode":           s.options.RuntimeConfig.SOCKS.Mode,
		"socks_listen_address": s.options.RuntimeConfig.SOCKS.ListenAddress,
	})
}

func (s *server) getConfig(w http.ResponseWriter, _ *http.Request) {
	writeJSON(w, http.StatusOK, s.options.RuntimeConfig)
}

func (s *server) saveConfig(w http.ResponseWriter, r *http.Request) {
	if strings.TrimSpace(s.options.ConfigPath) == "" {
		writeError(w, http.StatusServiceUnavailable, errors.New("configuration persistence is not enabled"))
		return
	}
	var candidate config.Config
	if err := decodeJSON(r, &candidate); err != nil {
		writeError(w, http.StatusBadRequest, err)
		return
	}
	if err := config.SaveAtomic(s.options.ConfigPath, candidate); err != nil {
		writeError(w, http.StatusBadRequest, err)
		return
	}
	writeJSON(w, http.StatusOK, map[string]any{
		"saved":            true,
		"restart_required": true,
		"config":           candidate,
	})
}

func (s *server) listLeases(w http.ResponseWriter, r *http.Request) {
	if r.URL.Query().Get("include_standby") == "true" {
		writeJSON(w, http.StatusOK, s.pool.ListAll())
		return
	}
	writeJSON(w, http.StatusOK, s.pool.List())
}

func (s *server) createLease(w http.ResponseWriter, r *http.Request) {
	var request createLeaseRequest
	if err := decodeJSON(r, &request); err != nil {
		writeError(w, http.StatusBadRequest, err)
		return
	}
	request.ID = strings.TrimSpace(request.ID)
	if request.ID == "" {
		writeError(w, http.StatusBadRequest, errors.New("id must not be empty"))
		return
	}

	s.mu.Lock()
	defer s.mu.Unlock()

	if s.options.RuntimeConfig.SOCKS.Mode != config.ModePerIPv6 || s.options.ListenerManager == nil {
		created, err := s.pool.Acquire(request.ID, request.Persistent)
		if err != nil {
			writeLeaseError(w, err)
			return
		}
		writeJSON(w, http.StatusCreated, created)
		return
	}

	created, existed := s.pool.Get(request.ID)
	if !existed {
		var err error
		created, err = s.pool.Acquire(request.ID, request.Persistent)
		if err != nil {
			writeLeaseError(w, err)
			return
		}
	}

	if _, running := s.options.ListenerManager.Get(request.ID); !running {
		host, _, err := net.SplitHostPort(s.options.RuntimeConfig.SOCKS.ListenAddress)
		if err != nil {
			if !existed {
				s.pool.Release(request.ID)
			}
			writeError(w, http.StatusInternalServerError, fmt.Errorf("parse SOCKS listen address: %w", err))
			return
		}
		address := net.JoinHostPort(host, strconv.Itoa(created.Port))
		if _, err := s.options.ListenerManager.Start(request.ID, address); err != nil {
			if !existed {
				s.pool.Release(request.ID)
			}
			writeError(w, http.StatusInternalServerError, err)
			return
		}
	}

	writeJSON(w, http.StatusCreated, created)
}

func writeLeaseError(w http.ResponseWriter, err error) {
	status := http.StatusInternalServerError
	switch {
	case errors.Is(err, lease.ErrCapacity), errors.Is(err, lease.ErrPortUnavailable):
		status = http.StatusConflict
	case errors.Is(err, lease.ErrNotFound):
		status = http.StatusNotFound
	case errors.Is(err, lease.ErrPersistent):
		status = http.StatusBadRequest
	}
	writeError(w, status, err)
}

func (s *server) getLease(w http.ResponseWriter, r *http.Request) {
	entry, ok := s.pool.Get(r.PathValue("id"))
	if !ok {
		writeError(w, http.StatusNotFound, errors.New("lease not found"))
		return
	}
	writeJSON(w, http.StatusOK, entry)
}

func (s *server) updateLease(w http.ResponseWriter, r *http.Request) {
	var request updateLeaseRequest
	if err := decodeJSON(r, &request); err != nil {
		writeError(w, http.StatusBadRequest, err)
		return
	}
	if request.Persistent == nil {
		writeError(w, http.StatusBadRequest, errors.New("persistent is required"))
		return
	}
	updated, err := s.pool.SetPersistent(r.PathValue("id"), *request.Persistent)
	if err != nil {
		writeError(w, http.StatusNotFound, err)
		return
	}
	writeJSON(w, http.StatusOK, updated)
}

func (s *server) deleteLease(w http.ResponseWriter, r *http.Request) {
	id := r.PathValue("id")

	s.mu.Lock()
	defer s.mu.Unlock()

	if _, exists := s.pool.Get(id); !exists {
		writeError(w, http.StatusNotFound, errors.New("lease not found"))
		return
	}
	if s.options.RuntimeConfig.SOCKS.Mode == config.ModePerIPv6 && s.options.ListenerManager != nil {
		if err := s.options.ListenerManager.Stop(id); err != nil {
			writeError(w, http.StatusInternalServerError, err)
			return
		}
	}
	if !s.pool.Release(id) {
		writeError(w, http.StatusNotFound, errors.New("lease not found"))
		return
	}
	w.WriteHeader(http.StatusNoContent)
}

func (s *server) releaseIdle(w http.ResponseWriter, _ *http.Request) {
	writeJSON(w, http.StatusOK, map[string]int{"released": s.pool.ReleaseIdle()})
}

// rotateLease assigns the lease a fresh IPv6 source address. The port stays
// unchanged in per_ipv6 mode, so clients only need to update any hardcoded IP
// allowlist, not their proxy configuration.
func (s *server) rotateLease(w http.ResponseWriter, r *http.Request) {
	s.mu.Lock()
	defer s.mu.Unlock()

	rotated, err := s.pool.Rotate(r.PathValue("id"))
	if err != nil {
		writeError(w, http.StatusNotFound, err)
		return
	}
	writeJSON(w, http.StatusOK, rotated)
}

// recycleLease releases the lease and immediately re-acquires a new one under
// the same id: the client keeps its identity but gets a different port and a
// fresh IPv6. In per_ipv6 mode the old listener is stopped by the OnRelease
// hook and a new one is started for the replacement lease.
func (s *server) recycleLease(w http.ResponseWriter, r *http.Request) {
	id := r.PathValue("id")

	s.mu.Lock()
	defer s.mu.Unlock()

	recycled, err := s.pool.Recycle(id)
	if err != nil {
		writeLeaseError(w, err)
		return
	}

	if s.options.RuntimeConfig.SOCKS.Mode == config.ModePerIPv6 && s.options.ListenerManager != nil {
		host, _, err := net.SplitHostPort(s.options.RuntimeConfig.SOCKS.ListenAddress)
		if err != nil {
			writeError(w, http.StatusInternalServerError, fmt.Errorf("parse SOCKS listen address: %w", err))
			return
		}
		address := net.JoinHostPort(host, strconv.Itoa(recycled.Port))
		if _, err := s.options.ListenerManager.Start(id, address); err != nil {
			writeError(w, http.StatusInternalServerError, err)
			return
		}
	}

	writeJSON(w, http.StatusOK, recycled)
}

func decodeJSON(r *http.Request, destination any) error {
	decoder := json.NewDecoder(http.MaxBytesReader(nil, r.Body, 1<<20))
	decoder.DisallowUnknownFields()
	if err := decoder.Decode(destination); err != nil {
		return err
	}
	var trailing any
	if err := decoder.Decode(&trailing); err == nil {
		return errors.New("request body must contain one JSON value")
	}
	return nil
}

func writeJSON(w http.ResponseWriter, status int, value any) {
	w.Header().Set("Content-Type", "application/json")
	w.WriteHeader(status)
	_ = json.NewEncoder(w).Encode(value)
}

func writeError(w http.ResponseWriter, status int, err error) {
	writeJSON(w, status, map[string]string{"error": err.Error()})
}